requirements.txt
fastapi==0.111.*
uvicorn==0.30.*
gunicorn==22.*
fastapi-users==14.*
fastapi-users[sqlalchemy]==14.*
python-dotenv==1.*
sqlalchemy==2.*
asyncpg==0.*
aiosqlite==0.*
pydantic==2.*
opcua==0.98.13
httpx==0.27.*
aiosmtpd==1.*
